Refactor: Move classes to 'models' package for clarity
Relocated 'SubstitutionRecord' and 'NutrientInfo' from 'services' to 'models' to better align with their purpose and structure. Adjusted imports across files to reflect the package changes. Simplified the 'Main' class by removing it from the 'utility' package.
