+++ src/models/SubstitutionRecord.java
@@ -1,4 +1,4 @@
-package services;
+package models;
 
 import java.time.LocalDate;
